LEAD packaging/Network-Assignment-lead-packaging/requirements.txt
python-dotenv==0.19.2
numpy==1.19.1
pandas==1.0.5
scipy==1.5.0
pyshp==2.1.0
shapely==1.7.0
numba==0.53.0
